Add node sub verb
