+++ .github/workflows/jazzy.yaml
@@ -33,7 +33,7 @@ jobs:
             auto_apms_mission
             auto_apms_simulation
             auto_apms_examples
-            auto_apms_ros2cli
+            auto_apms_ros2behavior
           import-token: ${{ secrets.GITHUB_TOKEN }}
           target-ros2-distro: jazzy
           vcs-repo-file-url: ${GITHUB_WORKSPACE}/dependencies.repos
