+++ auto_apms_examples/package.xml
@@ -16,7 +16,7 @@
     <depend>auto_apms_behavior_tree</depend>
     <depend>auto_apms_mission</depend>
     <depend>auto_apms_simulation</depend>
-    <depend>auto_apms_ros2cli</depend>
+    <depend>auto_apms_ros2behavior</depend>
 
     <!-- auto_apms_px4 must be installed manually if the examples should be generated -->
