+++ auto_apms_ros2behavior/auto_apms_ros2behavior/verb/node/manifests.py
@@ -1,0 +1,33 @@
+# Copyright 2025 Robin Müller
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from auto_apms_behavior_tree_core.resources import NodeManifest, get_node_manifest_resource_identities
+from ...verb import VerbExtension
+
+
+class ManifestsVerb(VerbExtension):
+    """List all registered behavior tree node manifests."""
+
+    def add_arguments(self, parser, cli_name):
+        pass
+
+    def main(self, *, args):
+        identities = get_node_manifest_resource_identities()
+        for i in identities:
+            print(str(i))
+            manifest = NodeManifest.from_resource(i)
+            for name in manifest.get_node_names():
+                print(f"  - {name} ({manifest.get_node_registration_options(name)["class_name"]})")
+        return 0
