--- auto_apms_ros2cli/resource/auto_apms_ros2cli
@@ -1,1 +1,0 @@
-auto_apms_ros2cli
