+++ auto_apms_util/auto_apms_util/resources.py
@@ -16,6 +16,8 @@ import os
 import ament_index_python
 import xml.etree.ElementTree as ET
 
+from collections import defaultdict
+
 
 class ResourceError(Exception):
     """Exception raised when resource operations fail."""
@@ -124,7 +126,7 @@ def get_plugin_xml_path(package: str) -> str:
     return os.path.join(base_path, paths[0])
 
 
-def collect_plugin_xml_paths(exclude_packages: set[str] = None) -> list[str]:
+def collect_plugin_xml_paths(exclude_packages: set[str] = None) -> list[tuple[str, str]]:
     """
     Collect the paths of plugins.xml manifest files generated by AutoAPMS.
 
@@ -135,7 +137,7 @@ def collect_plugin_xml_paths(exclude_packages: set[str] = None) -> list[str]:
         exclude_packages: Packages to exclude when searching for plugins.xml file paths.
 
     Returns:
-        List of absolute paths to the pluginlib-style plugin manifest xml files.
+        List of tuples with package name and absolute path to the pluginlib-style plugin manifest xml files.
 
     Raises:
         ResourceError: If failed to find a plugin manifest file.
@@ -143,12 +145,14 @@ def collect_plugin_xml_paths(exclude_packages: set[str] = None) -> list[str]:
     """
     xml_paths = []
     packages = get_packages_with_resource_type(_AUTO_APMS_UTIL__RESOURCE_TYPE_NAME__PLUGINLIB, exclude_packages)
-    for package in packages:
-        xml_paths.append(get_plugin_xml_path(package))
+    for p in packages:
+        xml_paths.append((p, get_plugin_xml_path(p)))
     return xml_paths
 
 
-def get_plugin_names_with_base_type(base_class_type: str, exclude_packages: set[str] = None) -> list[str]:
+def get_plugin_names_with_base_type(
+    base_class_type: str, exclude_packages: set[str] = None, *, per_package=False
+) -> list[str] | dict[str, list[str]]:
     """
     Find all plugin names that have a specific base class type from plugin XML manifest files.
 
@@ -158,17 +162,16 @@ def get_plugin_names_with_base_type(base_class_type: str, exclude_packages: set[
     Args:
         base_class_type: Fully qualified name of the base class type to search for.
         exclude_packages: Packages to exclude when searching for plugins.
+        per_package: If True, returns a dictionary with package names as keys and lists of plugin names as values.
 
     Returns:
-        List of plugin names (class names) that have the specified base class type.
+        Plugin names (fully qualified class names) that have the specified base class type.
 
     Raises:
         ResourceError: If failed to find or parse a plugin manifest file.
     """
-    plugin_names = []
-    xml_paths = collect_plugin_xml_paths(exclude_packages)
-
-    for xml_path in xml_paths:
+    plugins_per_package = defaultdict(list)
+    for package, xml_path in collect_plugin_xml_paths(exclude_packages):
         try:
             # Read the file content and preprocess it to fix XML encoding issues
             with open(xml_path, "r", encoding="utf-8") as f:
@@ -199,11 +202,15 @@ def get_plugin_names_with_base_type(base_class_type: str, exclude_packages: set[
                     if class_elem.get("base_class_type") == base_class_type:
                         plugin_name = class_elem.get("name")
                         if plugin_name:
-                            plugin_names.append(plugin_name)
+                            plugins_per_package[package].append(plugin_name)
 
         except ET.ParseError as e:
             raise ResourceError(f"Failed to parse plugin XML manifest file '{xml_path}': {e}")
         except Exception as e:
             raise ResourceError(f"Error reading plugin XML manifest file '{xml_path}': {e}")
 
-    return plugin_names
+    return (
+        plugins_per_package
+        if per_package
+        else [plugin_name for plugins in plugins_per_package.values() for plugin_name in plugins]
+    )
