+++ auto_apms_util/include/auto_apms_util/action_wrapper.hpp
@@ -178,8 +178,6 @@ ActionWrapper<ActionT>::ActionWrapper(
   std::shared_ptr<ActionContextType> action_context_ptr)
 : node_ptr_(node_ptr), action_context_ptr_(action_context_ptr), param_listener_(node_ptr)
 {
-  exposeToGlobalDebugLogging(node_ptr_->get_logger());
-
   using namespace std::placeholders;
   action_server_ptr_ = rclcpp_action::create_server<ActionT>(
     node_ptr_, action_name, std::bind(&ActionWrapper<ActionT>::handle_goal_, this, _1, _2),
